+++ lib/ruby-push-notifications/apns/apns_pusher.rb
@@ -16,6 +16,7 @@ module RubyPushNotifications
       # @param certificate [String]. The PEM encoded APNS certificate.
       # @param sandbox [Boolean]. Whether the certificate is an APNS sandbox or not.
       # @param options [Hash] optional. Options for APNSPusher. Currently supports:
+      #   * apns_url [String]: The URL of the APNS environment. Defaults to the official APNS URL.
       #   * connect_timeout [Integer]: Number of seconds to wait for the connection to open. Defaults to 30.
       def initialize(certificate, sandbox, password = nil, options = {})
         @certificate = certificate